Interactive 3D projection
The project has been updated to draw the plot of the dynamical system
using a 3D perspective projection onto the HTML canvas, instead of the
previous orthogonal projection. The user can orbit around the plot
interactively by clicking and dragging the mouse over the canvas. The
mouse wheel moves the eye closer/further.

Mobile and tablets have not been tested.

All the mathematics to perform the 3D projection is written in plain
JavaScript to demonstrate the basics of this technique. Because this is
a "from scratch" example for beginners, no specialised buffers or WebGL
were used and no effort was spent on optimisations other than culling
points that are too close to be resolved to distinct pixels.

Still, on Chrome on my laptop it is fast enough to maintain a real-time
feel up to about 50000 curve points from 500000 data points. Firefox is
slightly slower.